Scaffold for pxblue components and page templates
