+++ src/components/px-blue/data-display/Hero.tsx
@@ -1,0 +1,51 @@
+import React from 'react';
+import { useTheme, Theme, Typography } from '@material-ui/core';
+import { createStyles, makeStyles } from '@material-ui/styles';
+
+const useStyles = makeStyles((theme: Theme) =>
+    createStyles({
+        container: {
+            marginBottom: theme.spacing(4),
+        },
+        sectionTitle: {
+            marginBottom: theme.spacing(2),
+        },
+        exampleRow: {
+            display: 'flex',
+            justifyContent: 'space-evenly',
+            marginBottom: theme.spacing(2),
+        },
+        label: {
+            marginBottom: theme.spacing(1),
+        },
+        componentContainer: {
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'center',
+            flex: 1,
+        },
+    })
+);
+
+export const HeroExample: React.FC = () => {
+    const theme = useTheme();
+    const classes = useStyles(theme);
+
+    return (
+        <>
+            <div className={classes.container}>
+                <Typography variant={'body1'} className={classes.sectionTitle}>
+                    Section Title
+                </Typography>
+                <div className={classes.exampleRow}>
+                    <div className={classes.componentContainer}>
+                        <Typography variant={'body2'} className={classes.label}>
+                            Label
+                        </Typography>
+                        {/* component example goes here */}
+                    </div>
+                </div>
+            </div>
+        </>
+    );
+};
