+++ src/pages/index.tsx
@@ -1,3 +1,3 @@
 export * from './contextual-page-templates';
 export * from './material-ui';
-export * from './PXBlueComponents';
+export * from './px-blue';
